rodinia: fix memory leaks and uninitialized values in particlefilter
ParticleFilter in Rodinia fails when run on the head of GPGPU-Sim
with a memory corruption.  Looking further into the failures with
valgrind, there were numerous uninitialized values in the arrays
that were being accessed subsequently, causing the memory
corruption error.  To resolve this, all mallocs have been changed
to callocs, which intiializes all locations to 0 and removes the
error.  From looking at the code, it appears these locations
were assumed to be 0 all along, so this fix does not affect any
functionality.